working on shell out to docker
