+++ script/hello.js
@@ -1,0 +1,6 @@
+exports.handler = async function (event) {
+
+  return {statusCode: 200, body: '{"okay": true}'}
+}
+
+//SYNTAX ERROR
